# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for your tool-based Go backend.",
        "title": "Fiber Tools API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
        "/tools/age": {
            "get": {
                "description": "Takes birth year and returns current age",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tools"
                ],
                "summary": "Calculate age from birth year",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Birth year",
                        "name": "year",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tools/emicalculator": {
            "get": {
                "description": "Takes principal, rate, and tenure to calculate monthly EMI",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tools"
                ],
                "summary": "Calculate EMI (Equated Monthly Installment)",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Principal amount",
                        "name": "principal",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Annual interest rate in percent",
                        "name": "rate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Loan tenure in months",
                        "name": "tenure",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "number"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tools/passwordgenerator": {
            "get": {
                "description": "Generate a random password of specified length",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tools"
                ],
                "summary": "Generate random password",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Password length (default: 12)",
                        "name": "length",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tools/sipcalculator": {
            "get": {
                "description": "Takes monthly amount, rate, and months to calculate accumulated corpus",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tools"
                ],
                "summary": "Calculate SIP (Systematic Investment Plan) corpus",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Monthly investment amount",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Annual interest rate in percent",
                        "name": "rate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Investment period in months",
                        "name": "months",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "number"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tools/unitconverter": {
            "get": {
                "description": "Convert values between length, weight, and temperature units",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tools"
                ],
                "summary": "Convert between different units",
                "parameters": [
                    {
                        "type": "number",
                        "description": "Value to convert",
                        "name": "value",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Source unit (km, m, kg, g, c, f)",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target unit (km, m, kg, g, c, f)",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tools/wordcounter": {
            "get": {
                "description": "Count the number of words in the provided text",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tools"
                ],
                "summary": "Count words in text",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Text to count words in",
                        "name": "text",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer"
                            }
                        }
                    }
                }
            }
        }
    }
}